+++ apps/api/src/contexts/organisation/application/services/team.service.ts
@@ -4,7 +4,7 @@ import {
     TeamSearchQuery,
     UpdateTeamPayload,
 } from '@intra/shared-kernel';
-import { Inject, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { IdentityUserService } from 'src/contexts/identity/application/services/identity-user.service';
 import { TeamMembershipDomain } from '../../domain/team-membership.domain';
 import { TeamDomain } from '../../domain/team.domain';
@@ -61,6 +61,10 @@ export class TeamService {
 
     async delete(id: number): Promise<void> {
         await this.getById(id);
+        const members = await this.listMembers(id);
+        if (members.length > 0) {
+            throw new BadRequestException('Team #' + id + ' cannot be deleted. It has ' + members.length + ' members assigned to it.');
+        }
         await this.teams.deleteById(id);
     }
